Split pkg/k8s/resource.go in to two files (and likewise for resource_test.go)
