+++ pkg/k8s/resource.go
@@ -2,77 +2,13 @@ package k8s
 
 import (
 	"bytes"
-	"encoding/json"
 	"fmt"
 	"io"
-	"io/ioutil"
-	"os"
-	"path/filepath"
 	"strings"
-	"text/template"
 
-	"github.com/Masterminds/sprig"
 	ms "github.com/mitchellh/mapstructure"
-	"github.com/pkg/errors"
 	"gopkg.in/yaml.v2"
-
-	"github.com/datawire/teleproxy/pkg/supervisor"
 )
-
-var READY = map[string]func(Resource) bool{
-	"": func(Resource) bool { return false },
-	"Deployment": func(r Resource) bool {
-		// NOTE - plombardi - (2019-05-20)
-		// a zero-sized deployment never gets status.readyReplicas and friends set by kubernetes deployment controller.
-		// this effectively short-circuits the wait.
-		//
-		// in the future it might be worth porting this change to StatefulSets, ReplicaSets and ReplicationControllers
-		if r.Spec().GetInt64("replicas") == 0 {
-			return true
-		}
-
-		return r.Status().GetInt64("readyReplicas") > 0
-	},
-	"Service": func(r Resource) bool {
-		return true
-	},
-	"Pod": func(r Resource) bool {
-		css := r.Status().GetMaps("containerStatuses")
-		var cs Map
-		for _, cs = range css {
-			if !cs.GetBool("ready") {
-				return false
-			}
-		}
-		return true
-	},
-	"Namespace": func(r Resource) bool {
-		return r.Status().GetString("phase") == "Active"
-	},
-	"ServiceAccount": func(r Resource) bool {
-		_, ok := r["secrets"]
-		return ok
-	},
-	"ClusterRole": func(r Resource) bool {
-		return true
-	},
-	"ClusterRoleBinding": func(r Resource) bool {
-		return true
-	},
-	"CustomResourceDefinition": func(r Resource) bool {
-		conditions := r.Status().GetMaps("conditions")
-		if len(conditions) > 0 {
-			last := conditions[len(conditions)-1]
-			if last["status"] == "True" {
-				return true
-			} else {
-				return false
-			}
-		} else {
-			return false
-		}
-	},
-}
 
 // Map is a YAML/JSON-ish map with some convenience methods on it.
 type Map map[string]interface{}
@@ -189,28 +125,6 @@ func (r Resource) Spec() Map {
 	return Map(r).GetMap("spec")
 }
 
-func (r Resource) ReadyImplemented() bool {
-	if r.Empty() {
-		return false
-	}
-	kind := r.Kind()
-	_, ok := READY[kind]
-	return ok
-}
-
-func (r Resource) Ready() bool {
-	if r.Empty() {
-		return false
-	}
-	kind := r.Kind()
-	f, ok := READY[kind]
-	if ok {
-		return f(r)
-	} else {
-		return true
-	}
-}
-
 type Metadata map[string]interface{}
 
 func (r Resource) Metadata() Metadata {
@@ -280,96 +194,11 @@ func fixupMap(obj map[interface{}]interface{}) map[string]interface{} {
 	return result
 }
 
+// NewResourceFromYaml takes a (already-parsed) untyped YAML
+// structure, and fixes it up to be JSON-compatible, and returns it as
+// a Resource.
 func NewResourceFromYaml(yaml map[interface{}]interface{}) Resource {
 	return Resource(fixupMap(yaml))
-}
-
-func isTemplate(input []byte) bool {
-	return strings.Contains(string(input), "@TEMPLATE@")
-}
-
-func builder(dir string) func(string) (string, error) {
-	return func(dockerfile string) (string, error) {
-		return image(dir, dockerfile)
-	}
-}
-
-func image(dir, dockerfile string) (string, error) {
-	var result string
-	errs := supervisor.Run("BLD", func(p *supervisor.Process) error {
-		iidfile, err := ioutil.TempFile("", "iid")
-		if err != nil {
-			return err
-		}
-		defer os.Remove(iidfile.Name())
-		err = iidfile.Close()
-		if err != nil {
-			return err
-		}
-
-		ctx := filepath.Dir(filepath.Join(dir, dockerfile))
-		cmd := p.Command("docker", "build", "-f", filepath.Base(dockerfile), ".", "--iidfile", iidfile.Name())
-		cmd.Dir = ctx
-		err = cmd.Run()
-		if err != nil {
-			return err
-		}
-		content, err := ioutil.ReadFile(iidfile.Name())
-		if err != nil {
-			return err
-		}
-		iid := strings.Split(strings.TrimSpace(string(content)), ":")[1]
-		short := iid[:12]
-
-		registry := strings.TrimSpace(os.Getenv("DOCKER_REGISTRY"))
-		if registry == "" {
-			return errors.Errorf("please set the DOCKER_REGISTRY environment variable")
-		}
-		tag := fmt.Sprintf("%s/%s", registry, short)
-
-		cmd = p.Command("docker", "tag", iid, tag)
-		err = cmd.Run()
-		if err != nil {
-			return err
-		}
-
-		result = tag
-
-		cmd = p.Command("docker", "push", tag)
-		return cmd.Run()
-	})
-	if len(errs) > 0 {
-		return "", errors.Errorf("errors building %s: %v", dockerfile, errs)
-	}
-	return result, nil
-}
-
-func ExpandResource(path string) (result []byte, err error) {
-	input, err := ioutil.ReadFile(path)
-	if err != nil {
-		return nil, fmt.Errorf("%s: %v", path, err)
-	}
-	if isTemplate(input) {
-		funcs := sprig.TxtFuncMap()
-		funcs["image"] = builder(filepath.Dir(path))
-		tmpl := template.New(filepath.Base(path)).Funcs(funcs)
-		_, err := tmpl.Parse(string(input))
-		if err != nil {
-			return nil, fmt.Errorf("%s: %v", path, err)
-		}
-
-		buf := bytes.NewBuffer(nil)
-		err = tmpl.ExecuteTemplate(buf, filepath.Base(path), nil)
-		if err != nil {
-			return nil, fmt.Errorf("%s: %v", path, err)
-		}
-
-		result = buf.Bytes()
-	} else {
-		result = input
-	}
-
-	return
 }
 
 func ParseResources(name, input string) (result []Resource, err error) {
@@ -390,67 +219,6 @@ func ParseResources(name, input string) (result []Resource, err error) {
 	}
 }
 
-func LoadResources(path string) (result []Resource, err error) {
-	var input []byte
-	input, err = ExpandResource(path)
-	if err != nil {
-		return
-	}
-	result, err = ParseResources(path, string(input))
-	return
-}
-
-func SaveResources(path string, resources []Resource) error {
-	output, err := MarshalResources(resources)
-	if err != nil {
-		return fmt.Errorf("%s: %v", path, err)
-	}
-	err = ioutil.WriteFile(path, output, 0644)
-	if err != nil {
-		return fmt.Errorf("%s: %v", path, err)
-	}
-	return nil
-}
-
-func WalkResources(filter func(name string) bool, roots ...string) (result []Resource, err error) {
-	for _, root := range roots {
-		err = filepath.Walk(root, func(path string, info os.FileInfo, err error) error {
-			if err != nil {
-				return err
-			}
-
-			if !info.IsDir() && filter(path) {
-				rsrcs, err := LoadResources(path)
-				if err != nil {
-					return err
-				} else {
-					result = append(result, rsrcs...)
-				}
-			}
-
-			return nil
-		})
-		if err != nil {
-			return
-		}
-	}
-
-	return
-}
-
-func MarshalResources(resources []Resource) ([]byte, error) {
-	buf := bytes.NewBuffer(nil)
-	e := yaml.NewEncoder(buf)
-	for _, r := range resources {
-		err := e.Encode(r)
-		if err != nil {
-			return nil, err
-		}
-	}
-	e.Close()
-	return buf.Bytes(), nil
-}
-
 //func (r *Resource) MarshalJSON() ([]byte, error) {
 //	return json.Marshal(r)
 //}
@@ -466,24 +234,3 @@ func MarshalResources(resources []Resource) ([]byte, error) {
 //func (r Resource) UnmarshalJSON(data []byte) error {
 //	return json.Unmarshal(data, &Resource{})
 //}
-
-func MarshalResource(resource Resource) ([]byte, error) {
-	return MarshalResources([]Resource{resource})
-}
-
-func MarshalResourcesJSON(resources []Resource) ([]byte, error) {
-	buf := bytes.NewBuffer(nil)
-	e := json.NewEncoder(buf)
-	for _, r := range resources {
-		err := e.Encode(r)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return buf.Bytes(), nil
-}
-
-func MarshalResourceJSON(resource Resource) ([]byte, error) {
-	return MarshalResourcesJSON([]Resource{resource})
-}
